Avoid races by letting buffers fill
Initially, by stopping j.tail() by setting complete to true, there was a race in the tests where stdout/err could have closed before we finished reading from there; this is, for instance, why we enforce a cooldown period when running schedules. Waiting for a second is a rather untidy way of allowing tests to complete before pulling the rug out of the bottom. But it works.